// src/app/modules/mktg/interfaces/srn.interface.ts
export interface ISRN {
    _id?: string;
    umid?: string;
    fullName?: any;
    avg?: string;
    q111?: number;
    q112?: number;
    q113?: number;
    q121?: number;
    q122?: number;
    q123?: number;
    q131?: number;
    q132?: number;
    q133?: number;
    q141?: number;
    q211?: number;
    q212?: number;
    q213?: number;
    q221?: number;
    q222?: number;
    q223?: number;
    q31?: string;
    q321?: number;
    q322?: number;
    q323?: number;
    q324?: number;
    q41?: string;
    q42?: string;
    q43?: string;
    q44?: string;
    surveyAt?: Date | string;
    comment?: string;
    createdAt?: Date | String;
    updatedAt?: Date | String;
}